requirements.txt
finnhub-python
pandas
numpy
argparse
requests
beautifulsoup4
readability-lxml
playwright
PyYAML
crawl4ai
lxml
asyncio-throttle
aiofiles
python-dateutil
python-dotenv